changed animation of how it work steps
