+++ src/Pages/Home/HowItWorks/index.jsx
@@ -1,4 +1,5 @@
-import React, { useRef, useEffect } from "react";
+import React, { useState, useRef, useEffect } from "react";
+import { throttle } from "lodash";
 
 import HowItWorksCard from "../../../Components/HowItWorksCard";
 import BenefitCard from "../../../Components/BenefitCard";
@@ -9,11 +10,35 @@ import Benefit from "../../../Constants/Benefit.js";
 import "./index.css";
 
 const HowItWorks = () => {
+  const [isVisible, setIsVisible] = useState(false);
+  const divRef = useRef(null);
+
+  useEffect(() => {
+    function handleScroll() {
+      const div = divRef.current;
+      if (div) {
+        const rect = div.getBoundingClientRect();
+        const windowHeight =
+          window.innerHeight || document.documentElement.clientHeight;
+
+        if (rect.top <= windowHeight && rect.bottom >= 0) {
+          setIsVisible(true);
+        } else {
+          setIsVisible(false);
+        }
+      }
+    }
+
+    window.addEventListener("scroll", handleScroll);
+    return () => {
+      window.removeEventListener("scroll", handleScroll);
+    };
+  }, []);
   return (
     <div className="how-it-works-container">
       <h1 className="how-it-works-title">How It Works</h1>
       <div className="how-it-works-inner">
-        <div className="how-it-works-cards-container">
+        <div className="how-it-works-cards-container" ref={divRef}>
           {HowItWorksSteps.map((step, idx) => (
             <HowItWorksCard
               icon={<img src={step.image} alt="" width="32px" height="32px" />}
@@ -21,6 +46,7 @@ const HowItWorks = () => {
               content={step.content}
               idx={idx + 1}
               key={step.title}
+              isVisible={isVisible}
             />
           ))}
         </div>
